+++ phpstan-baseline.neon
@@ -1,6 +1,16 @@
 parameters:
 	ignoreErrors:
 		-
+			message: "#^Call to an undefined static method PHPUnit\\\\Framework\\\\Assert\\:\\:assertStringContainsStringIgnoringLineEndings\\(\\)\\.$#"
+			count: 1
+			path: src/Assertion.php
+
+		-
+			message: "#^Call to an undefined static method PHPUnit\\\\Framework\\\\Assert\\:\\:assertStringEqualIgnoringLineEndings\\(\\)\\.$#"
+			count: 1
+			path: src/Assertion.php
+
+		-
 			message: "#^Class PHPUnit\\\\Framework\\\\Constraint\\\\IsEqual constructor invoked with 5 parameters, 1\\-4 required\\.$#"
 			count: 2
 			path: src/Assertion.php
